Updated readme.md. Moved the axiosTest.js page into a folder of the same name. Updated some of the devDependencies to new versions in package.json. Changed over the 2 class components to functional components (generic and calendar/index.js).
