+++ client/src/components/calendar/index.js
@@ -1,41 +1,40 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import './calendar.css';
-import moment from 'moment';
 const startDateMin = new Date().toISOString().split('T')[0];
-class Calendar extends Component {
-    state = {
+
+const Calendar = () => {
+    const [state, setState] = useState({
         startDate: '',
         endDate: '',
-    };
+    });
 
-    handleChange = (e) => {
-        this.setState({
+    const handleChange = (e) => {
+        setState({
             [e.target.name]: e.target.value,
         });
-    }
+    };
 
-    render() {
-        return (
-            <div>
-                <h1>Calendar</h1>
-                <h3>Arrival Date</h3>
-                <input type="date"
-                    id="startDate"
-                    name="startDate"
-                    min={startDateMin}
-                    className="date-picker"
-                    onChange={this.handleChange}
-                />
-                <h3>Departure Date</h3>
-                <input type="date"
-                    id="endDate"
-                    name="endDate"
-                    min={this.state.startDate}
-                    className="date-picker"
-                    onChange={this.handleChange}
-                />
-            </div >
-        );
-    }
-}
+    return (
+        <div>
+            <h1>Calendar</h1>
+            <h3>Arrival Date</h3>
+            <input type="date"
+                id="startDate"
+                name="startDate"
+                min={startDateMin}
+                className="date-picker"
+                onChange={handleChange}
+            />
+            <h3>Departure Date</h3>
+            <input type="date"
+                id="endDate"
+                name="endDate"
+                min={state.startDate}
+                className="date-picker"
+                onChange={handleChange}
+            />
+        </div >
+    );
+};
+
 export default Calendar;
