+++ package.json
@@ -20,13 +20,13 @@
         "chai": "^4.2.0",
         "chai-http": "^4.3.0",
         "concurrently": "^5.2.0",
-        "eslint": "^6.7.2",
+        "eslint": "^7.6.0",
         "eslint-config-standard": "^14.1.1",
-        "eslint-plugin-import": "^2.21.1",
-        "eslint-plugin-node": "^10.0.0",
+        "eslint-plugin-import": "^2.22.0",
+        "eslint-plugin-node": "^11.1.0",
         "eslint-plugin-promise": "^4.2.1",
-        "eslint-plugin-react": "^7.20.0",
-        "eslint-plugin-react-hooks": "^4.0.4",
+        "eslint-plugin-react": "^7.20.5",
+        "eslint-plugin-react-hooks": "^4.0.8",
         "eslint-plugin-standard": "^4.0.1",
         "mocha": "^8.0.1",
         "nodemon": "^2.0.4"
